[client] split config I/O functions into config_io.go
